Encontra a quantidade de processos distribuídos numa unidade judiciária.

Encontra a quantidade de processos distribuídos numa unidade judiciária.

cnj_quantidade(
  fim,
  inicio = 0,
  ano,
  segmento,
  uf,
  distribuidor,
  funcao,
  expr = "sp_vazio"
)

Arguments

fim
Qualquer inteiro que seguramente é superior ao número de processos distribuídos

inicio
Se você sabe que não pode haver menos que tanto, coloca esse número para reduzir o número de buscas.

ano
Indicar o ano em questão

segmento
Segmento do Judiciário

uf
Unidade federativa.

distribuidor
Código do distribuídor. Encontrar no data-raw.

funcao
Função a ser aplicada para baixar: esaj::download_cpopg, tjdft::baixar_processo, etc.

expr
Expressão a ser avaliada. No caso de São Paulo, eu criei uma função interna chamada `sp_vazio`, que verifica se o tamanho do arquivo é menor que 90Mb.

@details Essa função adota um procedimento heurístico para encontrar a quantidade de processos distribuídos numa unidade judiciária em um determinado ano. À moda de busca binária, ela vai iniciar requisições a partir da média entre o ínicio e o fim indicados por você. Ela realiza dez requisições por vez, pois pode haver uma sequência de processos não existentes. Como cada função para baixar processos retorna um objeto distinto, você deve indicar uma expressão que retornará verdadeiro para casos de processos não existentes. Para São Paulo, basta usar a função interna sp_vazio. Futuramente, incluirei outras expressões para outros TJs, de modo que este parâmetro será obsoleto.

Value

Quantidade máxima aproximada de processos distribuídos.